requirements.txt
fastapi
uvicorn
playwright
google-generativeai
python-dotenv
boto3
bcrypt
pyjwt